feat: reorganized and renamed files
